Hide config types from custom code (#37)
* chore: rename config.go to library.go

* Include common and generate custom config types

The goal of this changes is to simplify reading server configuration by:

1. Moving common configuration values into the library:
The purpose of this is to allow common configuration to be shared across
projects. This approach also allows this config to be effectively hidden
from custom code.

2. Generating downstream configuration values:
Downstream config varies by system and hence requires generation. This
change generates these values to avoid custom code providing them manually.

3. Introducing a function to read a config file:
The common.ReadConfig function takes a file, a common config struct and a
custom config struct and populates the structs with the yaml file value.

Closes: #3