Change callback to 'map' error rather than 'handle' error
The current approach that the framework takes to handle internal errors
is to have generated code call custom code (via a callback) and request
that it 'handle' the error. Handling the error involves:
1. Logging the internal error
2. Mapping the internal error into an appropriate http error

Logging an internal error is something that the framework already handles
and, in the general case, mapping an internal error into an http error
is something that could be handled automatically if such a mapping
existed.

This change does the following:
1. Introduce error types (defined in sysl)
2. Change the interaction between the framework and custom code to
support internal error handling.

Internal errors are now 'handled' by the framework. The framework will
automatically log the internal error, then request that custom code
map the error to an http error if desired. If the internal error is
one of the error types defined in sysl (above), then custom code can
return nil (ignoring any custom mapping) and the framework will return
the values as defined in the sysl definition.

An example of an error type defined in sysl is:

!type BusinessLogicError [~error]:
    http_status <: string [value = "500"]
    http_code <: string [value = "1001"]
    http_message <: string [value = "foo"]

Requires: sysl v0.42.0

Closes: #1