build: force makefile to use bash shell with safer defaults
Previously the build had a problem where the `run-arrai` recipe
could fail (e.g. if `arrai` command line tool was not found
on the PATH) but this failure wouldn't be noticed by make, and
the build would succeed with exit code 0.

To avoid this, force make to use the bash shell with safer
defaults set so it can detect failures if any shell commands
exit nonzero.

This is a hack as it forces a dependency on bash. It would
be cleaner to figure out how to redefine `run-arrai`.