Rework CI to run "make all" in container
Purpose: previously the CI build only ran
a subset of the targets defined in the project's
Makefile. This meant that not all targets were
getting exercised during CI. By running "make all"
we can exercise and test more of the build targets
during CI, notably the "gen" target that actually
runs the code generation scripts in codegen/arrai.

1. Define builder.Dockerfile for custom dev environment image

The image includes all dependencies required to succesfully
invoke sysl-go's Makefile:
* install sysl
* install arrai
* install goimports
* install golangci-lint

2. Define a custom github action that uses the
builder.Dockerfile container image to run make

3. Modify the existing github workflow CI
to use the new builder action to run "make all"

What's still messy:

The versions of sysl, arrai are pinned to specific
release versions that are latest at time of writing.
This isn't ideal, it'd be preferable if these
versions could be factored out as params, and there
was an easy way to say "Give me a builder environment
using latest version of sysl or version 123 of arrai"

The version of arrai to use is not defined in a single
place: there's a version defined in builder.Dockerfile
and another version defined in Dockerfile.

There's an awkward hack in place between the two
invocations of make inside our github action docker
container, if this isn't done then github fails to
even build the container in the latter step. This is
even more distressing as we don't even want to build
the container image again, we'd be delighted to re
use the same image we built earlier during the first
step when we invoked "make all".

The "make check-tidy" construct seems questionable -
it isn't clear why we're trying to check that files
that are version controlled haven't been modified.
This adds a dependency on having git installed and
a .git folder present for the Makefile to work. It
might be better to rework this in another way, but
I'm not sure what the point of it is (is it some kind
of regression test for generated code, where the
expected generated code is already tracked by
version control ? ).