autogen - add support for user customising array of grpc.DialOption
new autogen feature:
* added optional AdditionalGrpcDialOptions hook to core.Hooks
* added optional OverrideGrpcDialOptions hook to core.Hooks

These new hooks can be used to add to or completely override the
list of grpc.DialOption used when an auto generated service
creates a connection to a target downstream gRPC service.

This change breaks the older codegen/arrai code generation
because the signature of BuildDownstreamClients has changed to
inject hooks.

This change to sysl-go/core is also backwards-incompatible
for the newer codegen/arrai/auto aka "autogen" because the
signatures of BuildDownstreamClients and
BuildDownstreamGRPCClient have changed to inject hooks.