Introduce viper for reading env variables. (#90)
The change introduces github.com/spf13/viper for reading env variables in Sysl-Go. Viper is a very powerful api and provides a lot of functionality to read/write env properties across a set of different sources such as files and remote key value stores. It even allow to setup watches over the application configuration. The change intentionally dampens the power of the api by exposing only a limited set of functionality. The design on the other hand is flexible enough to add new features as needed. The changes included in the changeset:

1. Go wrapper api for exposing viper.
2. Unit test cases.
3. Some sample config files.
4. Changes in Sysl-go for consuming the config api.