autogen'd apps - unknown config keys are errors
1. Adjust config loading logic of sysl-go autogenerated applications
so that configuration keys with no corresponding configuration field
are regarded as errors. This respores the old behaviour when loading
config prior to v0.139.0, before the config reading implementation
was switched for viper.

2. Change special config key used to customise the prefix used when
loading config from environment variables. Previously this was
"config.envPrefix". Now it is "envPrefix". This is backwards
incompatible change, but we don't believe anyone is using this
functionality yet.

3. note there's the tricky case where we dont want "envPrefix" to
be regarded as an unexpected config key even though it doesn't get
mapped to any field in the config structure.

4. added a new smoke test to prove that we can start a generated
app and load a config value from an environment variable.
BEWARE: THIS ONLY WORKS IF THERE'S ALSO A VALUE FOR THE CONFIG
KEY IN THE CONFIG FILE.

C.f. Introduce viper for reading env variables. (#90)