Introduces new logging framework (#247)
* Introduces new logging framework

This commit introduces a new logging framework, replacing the existing approach
to logging.

This change is backwards incompatible in some areas, see below for further
information.

The existing logging approach was to use the pkg logger internally and
recommend that teams either configure this logger before bootstrapping if
required or use a log hook to route logs to another logger. The new logging
framework takes a flexible approach, allowing logging implementations to be
wrapped through a central logging interface.

The following backwards incompatible changes are included:
1. Request/response payload logging
2. SensitiveString
3. Log level gradations

1. Request/response payload logging
In the original design, request and response payloads were only logged if the
logger (Logrus at that stage) was set to a particular log level. When the pkg
logger was introduced, this implicit configuration approach was maintained in
much the same manner (i.e. logged if the pkg logger was set to verbose). In
introducing a generic approach to logging, an explicit approach to deciding
whether payloads were logged was introduced through an option in the config:
library > log > logPayload
Existing implementations will need to set this configuration value explicitly.

2. SensitiveString
A recent addition to the external configuration was the use of a
SensitiveString, a string which would replace its characters with asterisks
when stringified. This class was incorrectly included in the common package,
not the config package. As a result of including the ability to retrieve the
config from the context, this class created a cyclic dependency. Attempting
to resolve the issue by leaving this class within the common package would
not have addressed the underlaying problem.
Existing implementation will need to update this dependency
from: common.SensitiveString
to:   config.SensitiveString

3. Log level gradations
In the original design, Logrus log levels (and other Logrus config) could be
specified within the external configuration [1]. Logrus provides several log
level gradations (trace, debug, warn, etc.) whereas the new logging framework
only provides three: debug, info, error. In order to support existing
configuration files, Logrus log levels provided in the config are assigned
an appropriate log level within the new framework and a warning is printed to
indicate if a deprecated value is used. The reduction in the number of
gradations shouldn't present a problem in practice and a suitably determined
developer could work around this limitation with a custom logger if such
gradation is of necessity.

[1] Interestingly, this configuration appears to have never actually been
applied by Sysl-go. Instead, Sysl-go relyied on custom code to apply these
values from the configuration directly.

* Includes additional logging documentation

Additional documentation added to describe the type of events logged,
notable fields logged and additional documentation at the call-site
of the logger hook.

* Fixes persistent bootstrap logging message

This fix addresses the persistence of the bootstrap logging message
in instances where no external logger is set before bootstrapping.