Add support for array query params in client (#266)
This PR adds support for the default OpenAPI 3.0 method. The go standard library net/url package doesn't offer a simple way to achieve the default OpenAPI2.0 method(as it will url encode the commas if supplied as a single value)

The default method to represent duplicated query parameters in OpenAPI 2.0 is
duplicate=foo,bar
The default method to represent duplicated query parameters in OpenAPI 3.0 is
duplicate=foo,duplicate=bar

Reflection is used to enable Slices to be detected and handled correctly. We aren't able to perform basic type checking because Aliases in Sysl are translated into discrete types in sysl-go. This approach was taken to avoid the complexity of type reference traversal in sysl-go

This change WILL require the sysl-go library to be updated in lockstep.

* Add support for array query params in client

* Add test to demonstrate behaviour of net/url query library

* Add test to demonstrate query param url encoding

* Move type switching to arrai function

* Fix linting issues

* Refactor to call library for query param encoding

Refactors the codegen to call a library function, so that the go code is easier to change, and is unit tested.

* Replace t.Log with assertions