only generate client service if it has endpoints and fix checks of grpc-client (#372)
This PR ensures that applications that only contain gRPC types and are directly used by the main application will not be generated as a client. Previously, in this scenario, the generated type applications have the wrong go code.

FIXME: gRPC types that are directly used by the main application uses the generated types from protos directly. However, if a downstream is a gRPC app and it uses gRPC types, the types are aliased into the generated protos. We should remove the aliases but it is a breaking change to users.